Vote for 1 candidate at a time.
This simulates a slightly more realistic join flow. In a real network
we would likely have a flow like this:

1. Joining node connects to us and requests to be a candidate. We
   vote to accept it if we have no other current candidate.
2. Joining node does resource proof over the course of 5 minutes.
3. If the node passes resource proof we vote to add it to a new block.

This patch essentially skips step 2, and combines steps 1 and 3 into voting
to immediately add the oldest candidate that hasn't timed out to the next block.
This should fix issues whereby we have too many candidates appearing at the same
and jostling for position in blocks.